// src/components/keys/enums.ts
export enum ABKeyType {
	A = 'A',
	B = 'B',
}

export enum ShoulderKeyType {
	RIGHT_SHOULDER = 'RIGHT_SHOULDER',
	LEFT_SHOULDER = 'LEFT_SHOULDER',
}

export enum KeyMap {
	DPAD_UP_KEY = 'W',
	DPAD_LEFT_KEY = 'A',
	DPAD_DOWN_KEY = 'S',
	DPAD_RIGHT_KEY = 'D',
	START_KEY = 'F',
	SELECT_KEY = 'V',
	A_KEY = 'O',
	B_KEY = 'K',
	RIGHT_SHOULDER_KEY = 'I',
	LEFT_SHOULDER_KEY = 'R',
}